Removing file extraction stubs
There are a number of tools that can extract text from PDFs, but they’re
mostly based on the same underlying library, which is old, probably
unmaintained, and incompatible with Python 3. The feature doesn’t seem
to be worth further investigation for now. Revert the commit if you’re
interested in building upon it.